chore: move modals to modals folder
