+++ src/app/club/dashboard/manage/calendar/page.tsx
@@ -1,13 +1,15 @@
 import ClubInfo from "@/components/dashboard/club/calendar/molecules/ClubInfo";
 import ClubCalendar from "@/components/dashboard/shared/organisms/ClubCalendar";
-import NewScheduleForm from "@/components/dashboard/club/calendar/organisms/NewScheduleForm";
+import NewScheduleFormModal from "@/components/dashboard/club/calendar/modals/NewScheduleFormModal";
 
 export default function CalendarPage() {
   return (
     <>
       <ClubInfo />
       <ClubCalendar />
-      <NewScheduleForm />
+      <div className="m-0">
+        <NewScheduleFormModal />
+      </div>
     </>
   );
 }
