+++ src/app/club/dashboard/manage/disband/page.tsx
@@ -1,13 +1,15 @@
 import ClubList from "@/components/dashboard/club/disband/organisms/ClubList";
-import DisbandClubModal from "@/components/dashboard/club/disband/organisms/DisbandClubModal";
-import DisbandClubFormModal from "@/components/dashboard/club/disband/organisms/DisbandClubFormModal";
+import DisbandClubModal from "@/components/dashboard/club/disband/modals/DisbandClubModal";
+import DisbandClubFormModal from "@/components/dashboard/club/disband/modals/DisbandClubFormModal";
 
 export default function ClubDisbandPage() {
   return (
     <>
       <ClubList />
-      <DisbandClubModal />
-      <DisbandClubFormModal />
+      <div className="m-0">
+        <DisbandClubModal />
+        <DisbandClubFormModal />
+      </div>
     </>
   );
 }
