+++ src/app/club/dashboard/manage/member/page.tsx
@@ -1,9 +1,9 @@
 import MemberTitle from "@/components/dashboard/club/member/molecules/MemberTitle";
-import CancelApplicationModal from "@/components/dashboard/club/member/organisms/CancelApplicationModal";
+import CancelApplicationModal from "@/components/dashboard/club/member/modals/CancelApplicationModal";
 import MemberList from "@/components/dashboard/club/member/organisms/MemberList";
 import MemberSearch from "@/components/dashboard/club/member/organisms/MemberSearch";
-import NewMemberDetailModal from "@/components/dashboard/club/member/organisms/NewMemberDetailModal";
-import RevertCancelationModal from "@/components/dashboard/club/member/organisms/RevertCancelationModal";
+import NewMemberDetailModal from "@/components/dashboard/club/member/modals/NewMemberDetailModal";
+import RevertCancelationModal from "@/components/dashboard/club/member/modals/RevertCancelationModal";
 
 export default function ClubMemberPage() {
   return (
@@ -11,9 +11,11 @@ export default function ClubMemberPage() {
       <MemberTitle />
       <MemberSearch />
       <MemberList />
-      <NewMemberDetailModal />
-      <CancelApplicationModal />
-      <RevertCancelationModal />
+      <div className="m-0">
+        <NewMemberDetailModal />
+        <CancelApplicationModal />
+        <RevertCancelationModal />
+      </div>
     </>
   );
 }
