+++ src/app/club/dashboard/manage/report/[id]/page.tsx
@@ -6,8 +6,8 @@ import ReportViewer from "@/components/dashboard/club/report/organisms/ReportVie
 import ReportTitle from "@/components/dashboard/club/report/molecules/ReportTitle";
 import ClubInfoCard from "@/components/dashboard/club/common/ClubInfoCard";
 import NewReportForm from "@/components/dashboard/club/report/organisms/NewReportForm";
-import ReportSubmitSuccessModal from "@/components/dashboard/club/report/organisms/ReportSubmitSuccessModal";
-import ReportCancelModal from "@/components/dashboard/club/report/molecules/ReportCancelModal";
+import ReportSubmitSuccessModal from "@/components/dashboard/club/report/modals/ReportSubmitSuccessModal";
+import ReportCancelModal from "@/components/dashboard/club/report/modals/ReportCancelModal";
 
 export default function ClubReportDetailPage() {
   const status = useSearchParams().get("status");
@@ -34,8 +34,10 @@ export default function ClubReportDetailPage() {
           </div>
         </>
       )}
-      <ReportSubmitSuccessModal />
-      <ReportCancelModal />
+      <div className="m-0">
+        <ReportSubmitSuccessModal />
+        <ReportCancelModal />
+      </div>
     </>
   );
 }
