+++ src/components/dashboard/club/member/modals/NewMemberDetailModal.tsx
